Send decoded DMX from Arduino at most every 200 milliseconds
The ESP32 was drowing in MQTT messages, because the Arduino and Pi were
just publishing the DMX values as quickly as they came in.

Now, the Arduino records the latest DMX value for each channel, updated
every time a full DMX frame is received, and only publishes the latest
values every 500 milliseconds.

This means the ESP32 is no longer struggling to keep up with the colour
changes, but it drastically increases the chance of erroneous DMX values
slipping in, if/when the Arduino DMX shield misreads the DMX timings.