Send decoded DMX from Arduino every 50ms, rather than 200ms
Now that the Raspberry Pi is generating less MQTT chatter, we can up
the frequency of the decoded DMX dumps from the Arduino.

This helps the lights feel a bit more responsive, but still doesn’t
solve the problem of weird flashing if/when the DMX shield misreads
the DMX slots (sporadically putting values into the wrong channels).